Change user uid type to bigint

+++ db/migrate/001_create_users.rb
@@ -1,7 +1,7 @@
 class CreateUsers < ActiveRecord::Migration[5.1]
   def change
     create_table :users, force: true do |t|
-      t.integer :uid
+      t.bigint :uid
     end
   end
 end
